feat(schema)!: kuba.yaml format changed (#17)
BREAKING CHANGE

Old syntax was:

```yaml
mappings:
  - environment-variable: "FOO"
    value: "BAR"
```

New syntax is:

```yaml
env:
  - FOO
    value: "BAR"
```

The reasoning behind this is to make the config less verbose and
easier to read.

As long as you don't want or can upgrade your configuration files,
use the release v0.8.0 which still supports the old configuration
format.