feat(cache): add cache (#28)
This is completely optional.

Setting this to either `true` (default caching of `12h`),
`false` (disable completely, which should be the default),
a `number` (in seconds) or a `string` like `1d`, `2w`, `72h`, `2y`
will then store the values in a sqlite db in `.cache/kuba/db.sqlite`.

The reason behind this is,
that you might want to have your GCP secrets of your dayjob not stored/cached,
but you have a personal/hobby project,
where you want to cache your bitwarden secrets.