feat(cli): add `--command` flag to `run` command (#41)
Add optional `--command` flag to `run` subcommand.

This takes a string as command, e.g.

```sh
kuba run --command "echo \$SOME_SECRET"
```
Previously it was only possible to
pass env to a "program/application".

This spawns a shell (based off the `$SHELL` variable,
or falls back to `sh`) and executes the command.